requirements.txt
openai-whisper>=20231117
pydub>=0.25.1
torch>=2.0.0
numpy>=1.21.0
tqdm>=4.64.0